refactor in cats package
